sprint 2 feedback and refactor.
