+++ src/pages/LabelList/Label.tsx
@@ -1,21 +1,25 @@
 import styled from 'styled-components';
+import lightOrDark from '../../utils/lightOrDark';
 
 type LabelPropType = {
   bgColorCode: string;
   name: string;
+  thin?: boolean;
   // children: string | JSX.Element;
 };
 
 type WrapperProp = {
   bgColorCode: string;
   textColor: string;
+  thin?: boolean;
 };
 
 const Wrapper = styled.a<WrapperProp>`
   display: inline-block;
   padding: 0 10px;
-  line-height: 24px;
-  height: 24px;
+  line-height: ${({ thin }) => (thin ? '20px' : '24px')};
+  height: ${({ thin }) => (thin ? '20px' : '24px')};
+  margin-right: ${({ thin }) => thin && '4px'};
   background-color: ${({ bgColorCode }) => bgColorCode && `#${bgColorCode}`};
   border-radius: 16px;
   color: #fff;
@@ -26,21 +30,10 @@ const Wrapper = styled.a<WrapperProp>`
     bgColorCode === 'ffffff' && 'border: 1px solid #d0d7de'};
 `;
 
-const Label = ({ bgColorCode, name }: LabelPropType) => {
-  const lightOrDark = (bgColor = '000080') => {
-    const r = parseInt(bgColor.slice(0, 2), 16);
-    const g = parseInt(bgColor.slice(2, 4), 16);
-    const b = parseInt(bgColor.slice(4, 6), 16);
-    const hsp = r * 0.3 + g * 0.6 + b * 0.1;
-    if (hsp > 127.5) {
-      return 'black';
-    } else {
-      return 'white';
-    }
-  };
+const Label = ({ bgColorCode, name, thin }: LabelPropType) => {
   const textColor = lightOrDark(bgColorCode);
   return (
-    <Wrapper bgColorCode={bgColorCode} textColor={textColor}>
+    <Wrapper bgColorCode={bgColorCode} textColor={textColor} thin={thin}>
       {name}
     </Wrapper>
   );
